Image-to-Text-Conversion/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
natsort==7.1.1
Pillow==8.3.2
six==1.16.0
torch==1.5.0+cpu
torchvision==0.6.0+cpu
Werkzeug==2.0.1


